+++ src/tests.rs
@@ -37,13 +37,7 @@ fn mapped_atlas_layout_from_char_map_creates_correct_character_map_and_layout() 
 }
 
 #[test]
-#[cfg_attr(
-    all(
-        feature = "gizmos",
-        not(feature = "DO_NOT_USE_internal_tests_disable_gizmos")
-    ),
-    ignore
-)]
+#[cfg_attr(feature = "gizmos", ignore = "test cannot run with `gizmos` feature")]
 fn image_font_plugin_initialization() {
     let (mut app, handle) = initialize_app_with_example_font(ExampleFont::Monospace);